pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
  "mypy>=1.20.0",
  "pre-commit>=4.5.1",
  "ruff>=0.15.9"
]

[project]
name = "M14404"
version = "0.1.0"
description = "LU5740"
requires-python = ">=3.11"
dependencies = [
  "starlette>=0.37.2,<2.0.0",
  "uvicorn[standard]>=0.30.0,<2.0.0",
  "aiosqlite>=0.20.0,<0.30.0",
  "tortoise-orm>=1.1.7,<2.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "anyio>=4.0.0,<5.0.0",
  "httpx>=0.27.0,<0.30.0"
]

[tool.hatch.build.targets.wheel]
packages = ["M14404"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unreachable = true
pretty = true
show_error_codes = true
